requirements.txt
customtkinter>=5.0.3
Pillow>=9.0.0
pygame>=2.1.0
matplotlib>=3.5.0
requests>=2.26.0
numpy>=1.21.0
